+++ unit_test/container/darray_unit.c
@@ -487,7 +487,7 @@ ARC_UNIT_TEST(iterators_erase)
 
     for (i = 0; i < 10000; i++)
     {
-        unsigned end_idx = arc_darray_size(darray) - 1;
+        unsigned long end_idx = arc_darray_size(darray) - 1;
         ARC_ASSERT_INT_EQUAL(arc_darray_position(it, end_idx), ARC_SUCCESS);
         arc_darray_erase(it);
     }
